move avatar/header from separator into message wrapper
